chore: extract enums & interfaces
