+++ packages/desktop/lib/electron/enums/ledger-method.enum.ts
@@ -1,0 +1,4 @@
+export enum LedgerMethod {
+    GenerateEvmAddress = 'generate-evm-address',
+    SignEvmTransaction = 'sign-evm-transaction',
+}
