+++ packages/desktop/lib/electron/interfaces/ledger-process-message.interface.ts
@@ -1,0 +1,11 @@
+import { LedgerMethod } from '../enums/ledger-method.enum'
+
+export interface ILedgerProcessMessage {
+    error?: string | Error | unknown
+    data: ILedgerProcessMessageData
+}
+
+export interface ILedgerProcessMessageData {
+    method: LedgerMethod
+    parameters: (string | unknown)[]
+}
